AlaramAi/requirements.txt
python-telegram-bot==20.8
APScheduler==3.10.4
python-dotenv==1.0.1
requests==2.31.0
twilio==9.0.1
pymongo[srv]==4.7.2
dnspython==2.6.1
